// src/app/types/patient.ts
export interface Patient {
    user_id: string;  
    first_name: string;
    last_name: string;
    date_of_birth: string;
    sex: string;
    phone: string;
    age: string;
    height: string;
    weight: string;
    password: string;
    bmi: string;
    blood_pressure: string;
    blood_type: string;
    resting_heart_rate: string;
    race: string;
    nationality: string;
    offspring: string;
  };